+++ modulo-2/projeto-whatslab/whatslabjaime-react-app/src/App.js
@@ -1,5 +1,4 @@
 // import logo from './logo.svg';
-import Mensageiro from './components/Mensageiro';
 import Header from './components/Header';
 import Main from './components/Main';
 import styled from 'styled-components';
@@ -7,8 +6,9 @@ import styled from 'styled-components';
 const ContainerApp = styled.div`
   display: flex;
   width: 46%;
-  height: 100vh;
-  border: 1px solid black;
+  height: 90vh;
+  border-radius: 25px;
+  //border: 1px solid black;
   margin: 0 auto;
   flex-direction: column;
   justify-content: space-between;
@@ -20,8 +20,7 @@ function App() {
     <ContainerApp>
       <Header/>
       <Main/>
-      <Mensageiro/> 
-    </ContainerApp>
+     </ContainerApp>
   );
 }
