+++ modulo-2/projeto-whatslab/whatslabjaime-react-app/src/components/Header.js
@@ -2,10 +2,12 @@ import React from 'react';
 import styled from 'styled-components'
 
 const ContainerHeader = styled.div`
-    background-color: gray;
+    background-color: #34af23;
+    border-radius: 25px;
     color: white;
     text-align: center;
     padding:10px;
+    margin-bottom: -20px;
 `
 
 class Header extends React.Component {
